Extract managed thread route services
